feat: unir controladores y rutas de auth en un solo achivo para cada topico
