--- src/routes/login.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const { loginController } = require('../controllers/')
-
-router.post('/', loginController.userAuth)
-
-module.exports = router;
