--- src/routes/signup.js
@@ -1,7 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const { signupController } = require('../controllers/')
-
-router.post('/', signupController.userSignup)
-
-module.exports = router;
